FAI-2350 - Vacancy Closed Handler error
Fix issue where handler was querying the live vacancy view type rather than the closed. I think this was done previously as the function app was receiving the event before the vacancy had been moved to the closed vacancy projection. Added polly to back off and retry then if not found throw an exception so it will be reprocessed by the calling function